decoder: Decouple schema & AST helpers into its own packages
