+++ decoder/decoder.go
@@ -6,9 +6,7 @@ package decoder
 import (
 	"fmt"
 
-	"github.com/hashicorp/hcl-lang/schema"
 	"github.com/hashicorp/hcl/v2"
-	"github.com/hashicorp/hcl/v2/hclsyntax"
 )
 
 type Decoder struct {
@@ -30,169 +28,6 @@ func posEqual(pos, other hcl.Pos) bool {
 	return pos.Line == other.Line &&
 		pos.Column == other.Column &&
 		pos.Byte == other.Byte
-}
-
-func mergeBlockBodySchemas(block *hcl.Block, blockSchema *schema.BlockSchema) (*schema.BodySchema, error) {
-	mergedSchema := &schema.BodySchema{}
-	if blockSchema.Body != nil {
-		mergedSchema = blockSchema.Body.Copy()
-	}
-	if mergedSchema.Attributes == nil {
-		mergedSchema.Attributes = make(map[string]*schema.AttributeSchema, 0)
-	}
-	if mergedSchema.Blocks == nil {
-		mergedSchema.Blocks = make(map[string]*schema.BlockSchema, 0)
-	}
-	if mergedSchema.TargetableAs == nil {
-		mergedSchema.TargetableAs = make([]*schema.Targetable, 0)
-	}
-	if mergedSchema.ImpliedOrigins == nil {
-		mergedSchema.ImpliedOrigins = make([]schema.ImpliedOrigin, 0)
-	}
-
-	depSchema, _, ok := NewBlockSchema(blockSchema).DependentBodySchema(block)
-	if ok {
-		for name, attr := range depSchema.Attributes {
-			if _, exists := mergedSchema.Attributes[name]; !exists {
-				mergedSchema.Attributes[name] = attr
-			} else {
-				// Skip duplicate attribute
-				continue
-			}
-		}
-		for bType, block := range depSchema.Blocks {
-			if _, exists := mergedSchema.Blocks[bType]; !exists {
-				// propagate DynamicBlocks extension to any nested blocks
-				if mergedSchema.Extensions != nil && mergedSchema.Extensions.DynamicBlocks {
-					if block.Body.Extensions == nil {
-						block.Body.Extensions = &schema.BodyExtensions{}
-					}
-					block.Body.Extensions.DynamicBlocks = true
-				}
-
-				mergedSchema.Blocks[bType] = block
-			} else {
-				// Skip duplicate block type
-				continue
-			}
-		}
-
-		if mergedSchema.Extensions != nil && mergedSchema.Extensions.DynamicBlocks && len(depSchema.Blocks) > 0 {
-			mergedSchema.Blocks["dynamic"] = buildDynamicBlockSchema(depSchema)
-		}
-
-		mergedSchema.TargetableAs = append(mergedSchema.TargetableAs, depSchema.TargetableAs...)
-		mergedSchema.ImpliedOrigins = append(mergedSchema.ImpliedOrigins, depSchema.ImpliedOrigins...)
-
-		// TODO: avoid resetting?
-		mergedSchema.Targets = depSchema.Targets.Copy()
-
-		// TODO: avoid resetting?
-		mergedSchema.DocsLink = depSchema.DocsLink.Copy()
-
-		// use extensions of DependentBody if not nil
-		// (to avoid resetting to nil)
-		if depSchema.Extensions != nil {
-			mergedSchema.Extensions = depSchema.Extensions.Copy()
-		}
-	} else if !ok && mergedSchema.Extensions != nil && mergedSchema.Extensions.DynamicBlocks && len(mergedSchema.Blocks) > 0 {
-		// dynamic blocks are only relevant for dependent schemas,
-		// but we may end up here because the schema is a result
-		// of merged static + dependent schema from previous iteration
-
-		// propagate DynamicBlocks extension to any nested blocks
-		if mergedSchema.Extensions != nil && mergedSchema.Extensions.DynamicBlocks {
-			for bType, block := range mergedSchema.Blocks {
-				if block.Body.Extensions == nil {
-					block.Body.Extensions = &schema.BodyExtensions{}
-				}
-				block.Body.Extensions.DynamicBlocks = true
-				mergedSchema.Blocks[bType] = block
-			}
-		}
-
-		mergedSchema.Blocks["dynamic"] = buildDynamicBlockSchema(mergedSchema)
-	}
-
-	return mergedSchema, nil
-}
-
-// blockContent represents HCL or JSON block content
-type blockContent struct {
-	*hcl.Block
-
-	// Range represents range of the block in HCL syntax
-	// or closest available representative range in JSON
-	Range hcl.Range
-}
-
-// bodyContent represents an HCL or JSON body content
-type bodyContent struct {
-	Attributes hcl.Attributes
-	Blocks     []*blockContent
-	RangePtr   *hcl.Range
-}
-
-// decodeBody produces content of either HCL or JSON body
-//
-// JSON body requires schema for decoding, empty bodyContent
-// is returned if nil schema is provided
-func decodeBody(body hcl.Body, bodySchema *schema.BodySchema) bodyContent {
-	content := bodyContent{
-		Attributes: make(hcl.Attributes, 0),
-		Blocks:     make([]*blockContent, 0),
-	}
-
-	// More common HCL syntax is processed directly (without schema)
-	// which also better represents the reality in symbol lookups
-	// i.e. expressions written as opposed to schema requirements
-	if hclBody, ok := body.(*hclsyntax.Body); ok {
-		for name, attr := range hclBody.Attributes {
-			content.Attributes[name] = attr.AsHCLAttribute()
-		}
-
-		for _, block := range hclBody.Blocks {
-			content.Blocks = append(content.Blocks, &blockContent{
-				Block: block.AsHCLBlock(),
-				Range: block.Range(),
-			})
-		}
-
-		content.RangePtr = hclBody.Range().Ptr()
-
-		return content
-	}
-
-	// JSON syntax cannot be decoded without schema as attributes
-	// and blocks are otherwise ambiguous
-	if bodySchema != nil {
-		hclSchema := bodySchema.ToHCLSchema()
-		bContent, remainingBody, _ := body.PartialContent(hclSchema)
-
-		content.Attributes = bContent.Attributes
-		if bodySchema.AnyAttribute != nil {
-			// Remaining unknown fields may also be blocks in JSON,
-			// but we blindly treat them as attributes here
-			// as we cannot do any better without upstream HCL changes.
-			remainingAttrs, _ := remainingBody.JustAttributes()
-			for name, attr := range remainingAttrs {
-				content.Attributes[name] = attr
-			}
-		}
-
-		for _, block := range bContent.Blocks {
-			// hcl.Block interface (as the only way of accessing block in JSON)
-			// does not come with Range for the block, so we calculate it here
-			rng := hcl.RangeBetween(block.DefRange, block.Body.MissingItemRange())
-
-			content.Blocks = append(content.Blocks, &blockContent{
-				Block: block,
-				Range: rng,
-			})
-		}
-	}
-
-	return content
 }
 
 func stringPos(pos hcl.Pos) string {
