+++ decoder/internal/schemahelper/dependent_body_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package decoder
+package schemahelper
 
 import (
 	"fmt"
